a few actions working
